+++ doc/ReleaseNotes.md
@@ -8,6 +8,7 @@ The release versions are NuGet releases.
 
 ### Fixes
 * fixed issue for `SvgTextBase.Bounds` with blank `SvgTextSpan` (see [PR #963](https://github.com/svg-net/SVG/pull/963))
+* fixed precision issue in `SvgUnit` (see [PR #969](https://github.com/svg-net/SVG/pull/969))
 
 ## [Version 3.4.1](https://www.nuget.org/packages/Svg/3.4.1)  (2022-03-19)